refactor: 공통 엔티티 common으로 이동
