+++ commerce/src/main/java/store/_0982/commerce/application/order/OrderPaymentProcessorService.java
@@ -11,9 +11,9 @@ import store._0982.commerce.domain.grouppurchase.GroupPurchase;
 import store._0982.commerce.domain.grouppurchase.GroupPurchaseRepository;
 import store._0982.commerce.domain.order.Order;
 import store._0982.commerce.domain.order.OrderRepository;
-import store._0982.commerce.domain.order.OrderStatus;
 import store._0982.commerce.domain.order.PaymentMethod;
 import store._0982.commerce.exception.CustomErrorCode;
+import store._0982.common.domain.order.OrderStatus;
 import store._0982.common.exception.CustomException;
 import store._0982.common.kafka.dto.PaymentChangedEvent;
 import store._0982.common.kafka.dto.PointChangedEvent;
