+++ commerce/src/main/java/store/_0982/commerce/application/product/dto/OrderVectorInfo.java
@@ -1,6 +1,6 @@
 package store._0982.commerce.application.product.dto;
 
-import store._0982.commerce.domain.order.OrderStatus;
+import store._0982.common.domain.order.OrderStatus;
 
 import java.time.OffsetDateTime;
 import java.util.UUID;
