+++ commerce/src/main/java/store/_0982/commerce/application/sellerbalance/SellerBalanceService.java
@@ -6,12 +6,15 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import store._0982.commerce.application.sellerbalance.dto.SellerBalanceCommand;
-import store._0982.commerce.application.sellerbalance.dto.SellerBalanceThumbnailInfo;
-import store._0982.commerce.domain.sellerbalance.*;
-import store._0982.commerce.exception.CustomErrorCode;
-import store._0982.common.dto.PageResponse;
 import store._0982.commerce.application.sellerbalance.dto.SellerBalanceHistoryInfo;
 import store._0982.commerce.application.sellerbalance.dto.SellerBalanceInfo;
+import store._0982.commerce.application.sellerbalance.dto.SellerBalanceThumbnailInfo;
+import store._0982.commerce.domain.sellerbalance.SellerBalanceHistoryRepository;
+import store._0982.commerce.domain.sellerbalance.SellerBalanceRepository;
+import store._0982.commerce.exception.CustomErrorCode;
+import store._0982.common.domain.sellerbalance.SellerBalance;
+import store._0982.common.domain.sellerbalance.SellerBalanceHistory;
+import store._0982.common.dto.PageResponse;
 import store._0982.common.exception.CustomException;
 
 import java.util.UUID;
