+++ commerce/src/main/java/store/_0982/commerce/application/sellerbalance/dto/SellerBalanceThumbnailInfo.java
@@ -1,6 +1,6 @@
 package store._0982.commerce.application.sellerbalance.dto;
 
-import store._0982.commerce.domain.sellerbalance.SellerBalance;
+import store._0982.common.domain.sellerbalance.SellerBalance;
 
 import java.util.UUID;
