+++ commerce/src/main/java/store/_0982/commerce/application/settlement/OrderSettlementService.java
@@ -5,9 +5,9 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import store._0982.commerce.domain.order.Order;
 import store._0982.commerce.domain.order.OrderCancellationPolicy;
-import store._0982.commerce.domain.order.OrderStatus;
-import store._0982.commerce.domain.settlement.OrderSettlement;
 import store._0982.commerce.domain.settlement.OrderSettlementRepository;
+import store._0982.common.domain.order.OrderStatus;
+import store._0982.common.domain.settlement.OrderSettlement;
 
 @Transactional(readOnly = true)
 @RequiredArgsConstructor
