+++ commerce/src/main/java/store/_0982/commerce/domain/order/OrderRepository.java
@@ -3,6 +3,7 @@ package store._0982.commerce.domain.order;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.repository.query.Param;
+import store._0982.common.domain.order.OrderStatus;
 
 import java.time.OffsetDateTime;
 import java.util.List;
