+++ commerce/src/main/java/store/_0982/commerce/domain/sellerbalance/SellerBalanceHistoryRepository.java
@@ -2,6 +2,7 @@ package store._0982.commerce.domain.sellerbalance;
 
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
+import store._0982.common.domain.sellerbalance.SellerBalanceHistory;
 
 import java.util.UUID;
