--- commerce/src/main/java/store/_0982/commerce/domain/sellerbalance/SellerBalanceHistoryStatus.java
@@ -1,6 +1,0 @@
-package store._0982.commerce.domain.sellerbalance;
-
-public enum SellerBalanceHistoryStatus {
-    CREDIT,     // 입금/증가
-    DEBIT       // 출금/감소
-}
