+++ commerce/src/main/java/store/_0982/commerce/domain/sellerbalance/SellerBalanceRepository.java
@@ -1,5 +1,7 @@
 package store._0982.commerce.domain.sellerbalance;
 
+import store._0982.common.domain.sellerbalance.SellerBalance;
+
 import java.util.Optional;
 import java.util.UUID;
