+++ commerce/src/main/java/store/_0982/commerce/domain/settlement/OrderSettlementRepository.java
@@ -1,5 +1,7 @@
 package store._0982.commerce.domain.settlement;
 
+import store._0982.common.domain.settlement.OrderSettlement;
+
 public interface OrderSettlementRepository {
 
     void save(OrderSettlement orderSettlement);
