+++ commerce/src/main/java/store/_0982/commerce/infrastructure/sellerbalance/SellerBalanceHistoryRepositoryAdapter.java
@@ -4,8 +4,8 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Repository;
-import store._0982.commerce.domain.sellerbalance.SellerBalanceHistory;
 import store._0982.commerce.domain.sellerbalance.SellerBalanceHistoryRepository;
+import store._0982.common.domain.sellerbalance.SellerBalanceHistory;
 
 import java.util.UUID;
