+++ commerce/src/main/java/store/_0982/commerce/infrastructure/sellerbalance/SellerBalanceJpaRepository.java
@@ -1,7 +1,7 @@
 package store._0982.commerce.infrastructure.sellerbalance;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import store._0982.commerce.domain.sellerbalance.SellerBalance;
+import store._0982.common.domain.sellerbalance.SellerBalance;
 
 import java.util.Optional;
 import java.util.UUID;
