+++ commerce/src/test/java/store/_0982/commerce/application/sellerbalance/SellerBalanceServiceTest.java
@@ -13,10 +13,7 @@ import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import store._0982.commerce.application.sellerbalance.dto.SellerBalanceHistoryInfo;
 import store._0982.commerce.application.sellerbalance.dto.SellerBalanceInfo;
-import store._0982.commerce.domain.sellerbalance.SellerBalance;
-import store._0982.commerce.domain.sellerbalance.SellerBalanceHistory;
 import store._0982.commerce.domain.sellerbalance.SellerBalanceHistoryRepository;
-import store._0982.commerce.domain.sellerbalance.SellerBalanceHistoryStatus;
 import store._0982.commerce.domain.sellerbalance.SellerBalanceRepository;
 import store._0982.common.dto.PageResponse;
 import store._0982.common.exception.CustomException;
