+++ common/src/main/java/store/_0982/common/domain/settlement/OrderSettlement.java
@@ -45,4 +45,28 @@ public class OrderSettlement {
 
     @Column(name = "settled_at")
     private OffsetDateTime settledAt;
+
+    public static OrderSettlement createOrderSettlement(
+            UUID orderId,
+            UUID sellerId,
+            UUID groupPurchaseId,
+            Long totalAmount,
+            OrderStatus orderStatus
+    ) {
+        return new OrderSettlement(orderId, sellerId, groupPurchaseId, totalAmount, orderStatus);
+    }
+
+    private OrderSettlement(
+            UUID orderId,
+            UUID sellerId,
+            UUID groupPurchaseId,
+            Long totalAmount,
+            OrderStatus orderStatus) {
+        this.orderSettlementId = UUID.randomUUID();
+        this.orderId = orderId;
+        this.sellerId = sellerId;
+        this.groupPurchaseId = groupPurchaseId;
+        this.totalAmount = totalAmount;
+        this.orderStatus = orderStatus;
+    }
 }
